face-ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

facenet-pytorch==2.5.3
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

numpy==1.26.4
Pillow==10.4.0
opencv-python==4.10.0.84